+++ src/main/java/org/example/matrix/Coordinate.java
@@ -1,4 +1,4 @@
 package org.example.matrix;
 
-public record Coordinate(int i, int j, long value) {
+public record Coordinate(int i, int j, double value) {
 }
